requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
msgpack==1.0.7
Pillow==10.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
rpds-py==0.16.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
zope.interface==6.1
